refactor(sidekick): split test for update root config
